+++ clients/privacy-center/cypress/e2e/consent-notices.cy.ts
@@ -76,11 +76,13 @@ describe("Privacy notice driven consent", () => {
     beforeEach(() => {
       cy.clearAllCookies();
       cy.getCookie(CONSENT_COOKIE_NAME).should("not.exist");
-      cy.visit("/consent");
-      cy.getByTestId("consent");
-      cy.overrideSettings(SETTINGS);
-      cy.getByTestId("consent");
-      cy.wait("@getVerificationConfig");
+      cy.visitConsent({
+        settingsOverride: {
+          IS_OVERLAY_ENABLED: true,
+          IS_GEOLOCATION_ENABLED: true,
+          GEOLOCATION_API_URL,
+        },
+      });
     });
 
     it("populates its header from the experience config", () => {
@@ -329,9 +331,8 @@ describe("Privacy notice driven consent", () => {
       };
       cy.setCookie(CONSENT_COOKIE_NAME, JSON.stringify(cookie));
       // Visit the consent page with notices enabled
-      cy.visit("/consent");
+      cy.visitConsent({ settingsOverride: SETTINGS });
       cy.getByTestId("consent");
-      cy.overrideSettings(SETTINGS);
       // Should follow state of consent cookie
       cy.wait("@getExperience").then(() => {
         cy.getByTestId(`consent-item-${PRIVACY_NOTICE_ID_1}`).within(() => {
@@ -358,9 +359,10 @@ describe("Privacy notice driven consent", () => {
 
   describe("consent reporting", () => {
     beforeEach(() => {
-      cy.visit("/consent");
+      cy.visitConsent({
+        settingsOverride: SETTINGS,
+      });
       cy.getByTestId("consent");
-      cy.overrideSettings(SETTINGS);
     });
 
     it("can make calls to consent reporting endpoints", () => {
@@ -391,8 +393,7 @@ describe("Privacy notice driven consent", () => {
     });
 
     it("renders hierarchical notices correctly", () => {
-      cy.visit("/consent");
-      cy.getByTestId("consent");
+      cy.visitConsent({ settingsOverride: SETTINGS });
       cy.overrideSettings(SETTINGS);
       cy.wait("@getExperience");
 
@@ -407,9 +408,7 @@ describe("Privacy notice driven consent", () => {
     });
 
     it("parent toggle should toggle all children", () => {
-      cy.visit("/consent");
-      cy.getByTestId("consent");
-      cy.overrideSettings(SETTINGS);
+      cy.visitConsent({ settingsOverride: SETTINGS });
       cy.wait("@getExperience");
 
       cy.getByTestId("consent-item-pri_notice-advertising-000")
@@ -442,9 +441,7 @@ describe("Privacy notice driven consent", () => {
     });
 
     it("toggle all children should toggle parent", () => {
-      cy.visit("/consent");
-      cy.getByTestId("consent");
-      cy.overrideSettings(SETTINGS);
+      cy.visitConsent({ settingsOverride: SETTINGS });
       cy.wait("@getExperience");
 
       cy.getByTestId("consent-item-pri_notice-advertising-000")
@@ -468,9 +465,7 @@ describe("Privacy notice driven consent", () => {
     });
 
     it("can save hierarchical notices", () => {
-      cy.visit("/consent");
-      cy.getByTestId("consent");
-      cy.overrideSettings(SETTINGS);
+      cy.visitConsent({ settingsOverride: SETTINGS });
       cy.wait("@getExperience");
 
       cy.getByTestId("consent-item-pri_notice-advertising-000").click();